# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Faren API - personalized profile platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      summary: Get user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /profile:
    get:
      operationId: getMyProfile
      tags: [profiles]
      summary: Get current user's profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile/links:
    post:
      operationId: addProfileLink
      tags: [profiles]
      summary: Add a social link to profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkBody"
      responses:
        "201":
          description: Link added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileLink"
  /profile/links/{linkId}:
    patch:
      operationId: updateProfileLink
      tags: [profiles]
      summary: Update a social link
      parameters:
        - name: linkId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLinkBody"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileLink"
    delete:
      operationId: deleteProfileLink
      tags: [profiles]
      summary: Delete a social link
      parameters:
        - name: linkId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /profile/discord:
    post:
      operationId: connectDiscord
      tags: [profiles]
      summary: Connect Discord account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectDiscordBody"
      responses:
        "200":
          description: Discord connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscordStatus"
    delete:
      operationId: disconnectDiscord
      tags: [profiles]
      summary: Disconnect Discord account
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        backgroundUrl:
          type: ["string", "null"]
        accentColor:
          type: ["string", "null"]
        backgroundOpacity:
          type: ["number", "null"]
        glowColor:
          type: ["string", "null"]
        cursorStyle:
          type: ["string", "null"]
        musicUrl:
          type: ["string", "null"]
        badges:
          type: array
          items:
            type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/ProfileLink"
        discordConnected:
          type: boolean
        discordUserId:
          type: ["string", "null"]
        discordUsername:
          type: ["string", "null"]
        discordAvatarUrl:
          type: ["string", "null"]
        discordStatus:
          type: ["string", "null"]
        musicConnected:
          type: boolean
        musicService:
          type: ["string", "null"]
        followersCount:
          type: integer
        followingCount:
          type: integer
        likesCount:
          type: integer
        viewsCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - username
        - badges
        - links
        - discordConnected
        - musicConnected
        - followersCount
        - followingCount
        - likesCount
        - viewsCount
        - createdAt
    PublicProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        backgroundUrl:
          type: ["string", "null"]
        accentColor:
          type: ["string", "null"]
        backgroundOpacity:
          type: ["number", "null"]
        glowColor:
          type: ["string", "null"]
        cursorStyle:
          type: ["string", "null"]
        musicUrl:
          type: ["string", "null"]
        badges:
          type: array
          items:
            type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/ProfileLink"
        discordConnected:
          type: boolean
        discordUsername:
          type: ["string", "null"]
        discordAvatarUrl:
          type: ["string", "null"]
        discordStatus:
          type: ["string", "null"]
        discordActivity:
          type: ["string", "null"]
        discordStatusEmoji:
          type: ["string", "null"]
        musicConnected:
          type: boolean
        musicService:
          type: ["string", "null"]
        nowPlaying:
          $ref: "#/components/schemas/NowPlaying"
        followersCount:
          type: integer
        followingCount:
          type: integer
        likesCount:
          type: integer
        viewsCount:
          type: integer
        isFollowing:
          type: boolean
        hasLiked:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - badges
        - links
        - discordConnected
        - musicConnected
        - followersCount
        - followingCount
        - likesCount
        - viewsCount
        - isFollowing
        - hasLiked
        - createdAt
    ProfileLink:
      type: object
      properties:
        id:
          type: integer
        platform:
          type: string
        label:
          type: string
        url:
          type: string
        iconUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - id
        - platform
        - label
        - url
        - sortOrder
    CreateLinkBody:
      type: object
      properties:
        platform:
          type: string
        label:
          type: string
        url:
          type: string
        iconUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - platform
        - label
        - url
    UpdateLinkBody:
      type: object
      properties:
        platform:
          type: string
        label:
          type: string
        url:
          type: string
        iconUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        backgroundUrl:
          type: ["string", "null"]
        accentColor:
          type: ["string", "null"]
        backgroundOpacity:
          type: ["number", "null"]
        glowColor:
          type: ["string", "null"]
        cursorStyle:
          type: ["string", "null"]
        musicUrl:
          type: ["string", "null"]
        badges:
          type: array
          items:
            type: string
    ConnectDiscordBody:
      type: object
      properties:
        discordUserId:
          type: string
        discordUsername:
          type: string
        discordAvatarUrl:
          type: ["string", "null"]
        discordStatus:
          type: ["string", "null"]
        discordActivity:
          type: ["string", "null"]
        discordStatusEmoji:
          type: ["string", "null"]
      required:
        - discordUserId
        - discordUsername
    DiscordStatus:
      type: object
      properties:
        connected:
          type: boolean
        discordUserId:
          type: ["string", "null"]
        discordUsername:
          type: ["string", "null"]
        discordAvatarUrl:
          type: ["string", "null"]
        discordStatus:
          type: ["string", "null"]
        discordActivity:
          type: ["string", "null"]
        discordStatusEmoji:
          type: ["string", "null"]
      required:
        - connected
...
    NowPlaying:
      type: object
      properties:
        isPlaying:
          type: boolean
        title:
          type: ["string", "null"]
        artist:
          type: ["string", "null"]
        album:
          type: ["string", "null"]
        albumArt:
          type: ["string", "null"]
        previewUrl:
          type: ["string", "null"]
        externalUrl:
          type: ["string", "null"]
        progress:
          type: ["integer", "null"]
        duration:
          type: ["integer", "null"]
      required:
        - isPlaying
...