# lib/api-zod/src/generated/types/discordStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Faren API - personalized profile platform
 * OpenAPI spec version: 0.1.0
 */

export interface DiscordStatus {
  connected: boolean;
  /** @nullable */
  discordUserId?: string | null;
  /** @nullable */
  discordUsername?: string | null;
  /** @nullable */
  discordAvatarUrl?: string | null;
  /** @nullable */
  discordStatus?: string | null;
  /** @nullable */
  discordActivity?: string | null;
  /** @nullable */
  discordStatusEmoji?: string | null;
}