# lib/api-zod/src/generated/types/nowPlaying.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Faren API - personalized profile platform
 * OpenAPI spec version: 0.1.0
 */

export interface NowPlaying {
  isPlaying: boolean;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  artist?: string | null;
  /** @nullable */
  album?: string | null;
  /** @nullable */
  albumArt?: string | null;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  externalUrl?: string | null;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  duration?: number | null;
}