# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Faren API - personalized profile platform
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileLink } from "./profileLink";

export interface Profile {
  id: number;
  userId: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  backgroundUrl?: string | null;
  /** @nullable */
  accentColor?: string | null;
  /** @nullable */
  backgroundOpacity?: number | null;
  /** @nullable */
  glowColor?: string | null;
  /** @nullable */
  cursorStyle?: string | null;
  /** @nullable */
  musicUrl?: string | null;
  badges: string[];
  links: ProfileLink[];
  discordConnected: boolean;
  /** @nullable */
  discordUserId?: string | null;
  /** @nullable */
  discordUsername?: string | null;
  /** @nullable */
  discordAvatarUrl?: string | null;
  /** @nullable */
  discordStatus?: string | null;
  musicConnected: boolean;
  /** @nullable */
  musicService?: string | null;
  followersCount: number;
  followingCount: number;
  likesCount: number;
  viewsCount: number;
  createdAt: string;
}