# lib/api-zod/src/generated/types/updateLinkBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Faren API - personalized profile platform
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateLinkBody {
  platform?: string;
  label?: string;
  url?: string;
  /** @nullable */
  iconUrl?: string | null;
  sortOrder?: number;
}